Allow Pathname objects for Stackprof :out
The code for `stackprof_results` currently only does checks to see if
the VALUE for `_stackprof.out` is a `T_STRING`, otherwise it assumes it
is an IO object.

This updates the code to first check if `Pathname` has been required and
defined.  If it is and the `_stackprof.out` is a `Pathname`, then open a
file in the same manner as if it was a string, but calls `.to_s` to
convert it to a ruby string first.

Fixes #79