Add `--alphabetical-flamegraph` for population-based instead of timeline
The current flamegraph of stackprof is often chopped because it puts the
passage of time on the x-axis.

This change adds `--alphabetical-flamegraph` mode which sorts the stack
traces in alphabetical order.  This view is much easier than the current
one for understanding bottlenecks.

Also, `--timeline-flamegraph` is added for consistency, and the original
`--flamegraph` option is the same effect as `--timeline-flamegraph` (for
compatibility reason).