Add optional IO arg for .print_stackcollapse
Like the other methods in the class, it allows for choosing a IO object
to have the data written to.  This can allow for writing to a IO object
in the current ruby process and then processing it in the same process
without having to do crazy STDOUT constant overwriting to achieve it.