Use rake tasks on travis
Will help keep the list of ruby versions found in the .travis.yml in
sync with the ones in the Rakefile, since if a new version of ruby is
added to the .travis.yml, but it will break the tasks.

This makes `rake test:docker:all` have a better chance of accurately
representing a full Travis build across all rubies.