[FIXUP] Final form
I think this is the route I am going to go with, and will most likely
rebase these [FIXUP] and [DEBUG] commits out.

It makes use of the `bundle install` that is already done by travis
anyway (since we did a `before_install` instead of overwriting the
`install`, this was being done when I switched the language back to
`ruby`), it makes for a simpler diff to digest.

Also... I think the `bundle exec rake ...` might be required...