Avoid unsafe mallocs inside signals
In 6c3f4d7c5234cb3274c4a624ea6cb52c26be3833 we started recording signals
immediately on receiving a signal, after observing that in Ruby 3.0+ the
CFP structure was safe to read. This resulted in much more accurate
measurements.

Unfortunately though rb_profile_frames seems to be safe to call inside a
signal handler, stackprof_record_sample_for_stack is not, since it calls
malloc (itself or through st_*) which is not async-signal-safe. This
could potentially result in deadlocks or memory corruption depending on
the malloc implementation.

This commit attempts to restore the safety of the postponed job approach
while keeping the accuracy of measuring immediately. Inside the signal
handler we record the frames to a buffer, then enqueue a postponed job
to do the actual recording.