Fix GC profiling timing
We were recording GC profile timestamps inside the VM postponed job that
flushes temporary information.  The postponed job could be run much
later than the sample was actually taken, so this commit records a
timestamp in the signal handler when there is a GC event.

It only records one time stamp, so if there are multiple GC events in a
row, we have to assume that the total duration spent in GC starts from
the first time stamp and extends until the timestamp of the next non-GC
sample.  In other words, we don't record a correct timestamp for each GC
sample, only the first one.